backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
motor
pydantic

# Testing
pytest
pytest-asyncio
httpx
anyio

